Response now a marker trait, OK handles its own json serialisation.
I've managed to add an implicit  def that itself accepts an implicit parameter of Writes[A] and returns Writes[OK[A]]. I am assuming that this allows the call site to tunnel an in-scope implicit into the machinery and it seems to work like a charm. I have learned that parameters must be in the case class constructor if they are to be picked up automatically but this isn't always going to be favourite so it's good to have the means to handle those cases.